feat(cmd): add `source` and `image` subcommands to `scan` (#1519)
**This change does not break any existing behaviour.**
- Creates `source` and `image` subcommands for `scan`.
- Inserts `source` as the default subcommand if none is provided.
- Removes the `experimental-oci-image` flag and its tests.
- Adds a feedback link to HTML output

For project scanning, users can use the following commands:
- `osv-scanner <file_name>`
- `osv-scanner scan <file_name>`
- `osv-scanner scan source <file_name>`

For docker scanning, users can use the following commands:
- `osv-scanner scan image <docker_image>`
- `osv-scanner scan image --archive <docker_image.tar>`

Help command:
```
NAME:
  osv-scanner - scans various mediums for dependencies and checks them against the OSV database

USAGE:
  osv-scanner [global options] command [command options]

EXAMPLES:
  # Scan a source directory
  $ osv-scanner scan source -r <source_directory>

  # Scan a container image
  $ osv-scanner scan image <image_name>

  # Scan a local image archive (e.g. a tar file) and generate HTML output
  $ osv-scanner scan image --serve --archive <image_name.tar>

  # Fix vulnerabilities in a manifest file and lockfile (non-interactive mode)
  $ osv-scanner fix --non-interactive -M <manifest_file> -L <lockfile>

  For full usage details, please refer to the help command of each subcommand (e.g. osv-scanner scan --help).

VERSION:
  1.9.1

COMMANDS:
  scan     scans projects and container images for dependencies, and checks them against the OSV database.
  fix      [EXPERIMENTAL] scans a manifest and/or lockfile for vulnerabilities and suggests changes for remediating them
  help, h  Shows a list of commands or help for one command


GLOBAL OPTIONS:
    --help, -h  show help  --version, -v  print the version

```
