+++ internal/output/table.go
@@ -100,6 +100,8 @@ func tableBuilder(outputTable table.Writer, vulnResult *models.VulnerabilityResu
 }
 
 func printContainerScanningResult(result Result, outputWriter io.Writer, terminalWidth int) {
+	// Add a newline to separate results from logs.
+	fmt.Fprintln(outputWriter)
 	fmt.Fprintf(outputWriter, "Container Scanning Result (%s):\n", result.ImageInfo.OS)
 	summary := fmt.Sprintf(
 		"Total %[1]d packages affected by %[2]d vulnerabilities (%[3]d Critical, %[4]d High, %[5]d Medium, %[6]d Low, %[7]d Unknown) from %[8]d ecosystems.\n"+
@@ -182,9 +184,9 @@ func printContainerScanningResult(result Result, outputWriter io.Writer, termina
 	fmt.Fprintln(outputWriter)
 
 	const promptMessage = "For the most comprehensive scan results, we recommend using the HTML output: " +
-		"`osv-scanner --format html --output results.html`.\n" +
+		"`osv-scanner scan image --serve <image_name>`.\n" +
 		"You can also view the full vulnerability list in your terminal with: " +
-		"`osv-scanner --format vertical`."
+		"`osv-scanner scan image --format vertical <image_name>`."
 	fmt.Fprintln(outputWriter, promptMessage)
 }
