+++ pkg/reporter/format.go
@@ -7,7 +7,7 @@ import (
 	"github.com/google/osv-scanner/pkg/models"
 )
 
-var format = []string{"table", "vertical", "json", "markdown", "sarif", "gh-annotations", "cyclonedx-1-4", "cyclonedx-1-5"}
+var format = []string{"table", "html", "vertical", "json", "markdown", "sarif", "gh-annotations", "cyclonedx-1-4", "cyclonedx-1-5"}
 
 func Format() []string {
 	return format
